Add a formative project which checks if the bank card is valid or invalid
The program uses an algorithm that multiplies every second number by two and then sums them up with the other numbers on the card.
If the modulus of the result is zero, it means that the card is correct.
Currently, the program can find every second number, multiply it, and close the loop after reaching the last number.